FNP-2 #added search results page
